Move DCE module to Language.PureScript.DCE
It will contian code which is unrealated to CoreFn.
